backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pymupdf==1.24.5
openai==1.35.0
sentence-transformers==3.0.0
chromadb==0.5.3
python-dotenv==1.0.1
